+++ reactjs/src/native/components/organisms/Home.tsx
@@ -7,7 +7,7 @@ import Users from '~/native/containers/UsersContainer';
 import Settings from '~/native/containers/SettingsContainer';
 import GeneralTopics from '~/native/components/molecules/GeneralTopics';
 import ProfileCategory from '~/native/containers/ProfileCategoryContainer';
-import AuthScreen from '../atoms/Auth';
+import Auth from '~/native/containers/AuthContainer';
 import Layout from '../templates/PaddingLayout';
 import Info from './Info';
 import Meigen from './Meigen';
@@ -23,7 +23,7 @@ const Home = (props: any) => {
         <Route path="/topics" render={props => <GeneralTopics {...props} />} />
         <Route path="/categories" render={props => <Categories {...props} />} />
         <Route path="/challenges" render={props => <Challenges {...props} />} />
-        <Route path="/login" render={props => <AuthScreen {...props} />} />
+        <Route path="/login" render={props => <Auth {...props} />} />
         <Route path="/info" render={props => <Info {...props} />} />
         <Route path="/meigen" render={props => <Meigen {...props} />} />
         <Route path="/contact" render={props => <Contact {...props} />} />
