+++ reactjs/src/web/components/atoms/Auth.tsx
@@ -3,10 +3,9 @@ import { Dialog, DialogContent } from '@material-ui/core';
 import DialogTitle, { DialogTitleProps } from '@material-ui/core/DialogTitle';
 import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
 import styled from 'styled-components';
-import shortid from 'shortid';
 import * as firebaseui from 'firebaseui';
+import firebase from '~/lib/firebase';
 
-import firebase, { uploadPhotoURLAsync } from '~/lib/firebase';
 import theme from '~/lib/theme';
 import { TITAN_TERMS_OF_USE, TITAN_PRIVACY_POLICY } from '~/constants/appInfo';
 
@@ -23,67 +22,7 @@ const StyledDialogTitle = styled(DialogTitle)`
 ` as React.ComponentType<DialogTitleProps>;
 
 const AuthModal = (props: any) => {
-  const { onClose, title, ...other } = props;
-
-  const signInSuccessWithAuthResult = (
-    credentials: firebase.auth.UserCredential
-  ) => {
-    const { user } = credentials;
-
-    const isTwitter =
-      credentials.additionalUserInfo &&
-      credentials.additionalUserInfo.providerId === 'twitter.com';
-
-    const data = {
-      id: user!.uid,
-      shortId: shortid.generate(),
-      displayName: user!.displayName,
-      photoURL: user!.photoURL,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-      twitterUsername: isTwitter
-        ? (credentials.additionalUserInfo! as any).username
-        : ''
-    };
-
-    const secureId = shortid.generate();
-    const dataSecure = {
-      id: secureId,
-      email: user!.email,
-      accessTokenKey: isTwitter
-        ? (credentials.credential! as any).accessToken
-        : '',
-      accessTokenSecret: isTwitter
-        ? (credentials.credential! as any).secret
-        : ''
-    };
-
-    const userRef = firebase
-      .firestore()
-      .collection('users')
-      // uidにしないと、reduxのprofileとfirestoreのusersが同期しない。
-      .doc(user!.uid);
-
-    userRef.get().then(doc => {
-      if (!doc.exists) {
-        userRef.set(data);
-        userRef
-          .collection('securities')
-          .doc(secureId)
-          .set(dataSecure)
-          .then(() => {
-            if (data.photoURL && data.photoURL !== '') {
-              uploadPhotoURLAsync(
-                data.photoURL,
-                data.shortId,
-                `/users/${data.id}`
-              );
-            }
-          });
-      }
-    });
-    return false;
-  };
+  const { onClose, title, signInSuccessWithAuthResult, ...other } = props;
 
   const uiConfig = {
     signInFlow: 'popup',
