+++ reactjs/src/web/containers/AuthContainer.tsx
@@ -1,0 +1,4 @@
+import connect from '~/connects/AuthConnect';
+import Auth from '~/web/components/atoms/Auth';
+
+export default connect(Auth);
